extra-requirements.txt
easydict
torch
torchvision
PyYAML
tqdm
ymir-exc